Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 
 
 int main() {
-    int num1, num2, num3, sum;
+    float num1, num2, num3, sum;
     float av;
     scanf("%d %d %d",&num1, &num2, &num3);
     //sum = num1+num2+num3;
